Change VTCP_Check() to take a ssize_t
Since the input value is sometimes the result of a read()/write() call,
avoid truncating the ssize_t value on calling it.

+++ include/vtcp.h
@@ -35,7 +35,7 @@ struct suckaddr;
 #define VTCP_ADDRBUFSIZE		64
 #define VTCP_PORTBUFSIZE		16
 
-int VTCP_Check(int a);
+int VTCP_Check(ssize_t a);
 #define VTCP_Assert(a) assert(VTCP_Check(a))
 
 struct suckaddr *VTCP_my_suckaddr(int sock);
